pyproject.toml
[tool.poetry]
name = "timetable"
version = "0.1.0"
description = ""
authors = ["Jerry <zhangji87@gmail.com>"]
readme = "README.md"

[tool.poetry.extras]
gunicorn = ["gunicorn"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.5"
sqlalchemy = "^2.0.17"
apispec = {version = "^5.2.2", extras = ["marshmallow"]}
apispec-webframeworks = "^0.5.2"
python-dateutil = "^2.8.2"
gunicorn = {version = "^20.1.0", optional = true}
setuptools = "^66.1.1"
python-dotenv = "^0.21.1"
pymysql = "^1.0.3"
flask-login = "^0.6.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
mypy = "^1.3.0"
types-python-dateutil = "^2.8.2"
ruff = "^0.0.267"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "default"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "W", "Q", "COM", "I", "N", "B", "RUF"]
line-length = 100
target-version = "py310"

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401"]

[tool.ruff.flake8-quotes]
inline-quotes = "single"
